du manager: fix ipv4 asn.1 to srsgnb type conversion

+++ lib/asn1/asn1_utils.cpp
@@ -922,8 +922,8 @@ uint64_t octet_string_helper::to_uint(const byte_buffer& buf)
 
 void octet_string_helper::to_octet_string(srsgnb::span<uint8_t> buf, uint64_t number)
 {
-  size_t nbytes = buf.size();
-  if ((1U << (8U * nbytes)) <= number) {
+  uint64_t nbytes = buf.size();
+  if ((static_cast<uint64_t>(1U) << (8U * nbytes)) <= number) {
     log_error("Integer={} does not fit in an OCTET STRING of size={}", number, nbytes);
     return;
   }
